+++ .ci.sh
@@ -583,143 +583,135 @@ shCiBuildWasm() {(set -e
     printf '' > .tmp.js
 	printf '
 /*jslint-disable*/
-var initSqlJsPromise = undefined;
-var initSqlJs = function (moduleConfig) {
-    if (initSqlJsPromise){
-      return initSqlJsPromise;
-    }
-    initSqlJsPromise = new Promise(function (resolveModule, reject) {
-        var Module = typeof moduleConfig !== "undefined" ? moduleConfig : {};
-        var originalOnAbortFunction = Module["onAbort"];
-        Module["onAbort"] = function (errorThatCausedAbort) {
-            reject(new Error(errorThatCausedAbort));
-            if (originalOnAbortFunction){
-              originalOnAbortFunction(errorThatCausedAbort);
-            }
-        };
-        Module["postRun"] = Module["postRun"] || [];
-        Module["postRun"].push(function () {
-            resolveModule(Module);
-        });
-        module = undefined;
+(function () {
+    "use strict";
 ' >> .tmp.js
     cat sqlmath_wasm.js | sed -e "s|/\*jslint-[a-z]*\*/||g" >> .tmp.js
     printf '
-        return Module;
-    }); // The end of the promise being returned
-  return initSqlJsPromise;
-} // The end of our initSqlJs function
-if (typeof exports === "object" && typeof module === "object"){
-    module.exports = initSqlJs;
-    module.exports.default = initSqlJs;
-}
-else if (typeof define === "function" && define["amd"]) {
-    define([], function() { return initSqlJs; });
-}
-else if (typeof exports === "object"){
-    exports["Module"] = initSqlJs;
-}
-/* global initSqlJs */
-/* eslint-env worker */
-/* eslint no-restricted-globals: ["error"] */
-
-"use strict";
-
-// hack-sqljs - conditional-worker
-if ((/\\binitSqlJsWorker=1\\b/).test(typeof location === "object" && location && location.search)) {
-var db;
-
-function onModuleReady(SQL) {
-    function createDb(data) {
-        if (db != null) db.close();
-        db = new SQL.Database(data);
-        return db;
-    }
-
-    var buff; var data; var result;
-    data = this["data"];
-    var config = data["config"] ? data["config"] : {};
-    switch (data && data["action"]) {
-        case "open":
-            buff = data["buffer"];
-            createDb(buff && new Uint8Array(buff));
-            return postMessage({
-                id: data["id"],
-                ready: true
-            });
-        case "exec":
-            if (db === null) {
-                createDb();
-            }
-            if (!data["sql"]) {
-                throw "exec: Missing query string";
-            }
-            return postMessage({
-                id: data["id"],
-                results: db.exec(data["sql"], data["params"], config)
-            });
-        case "each":
-            if (db === null) {
-                createDb();
-            }
-            var callback = function callback(row) {
-                return postMessage({
-                    id: data["id"],
-                    row: row,
-                    finished: false
-                });
-            };
-            var done = function done() {
-                return postMessage({
-                    id: data["id"],
-                    finished: true
-                });
-            };
-            return db.each(data["sql"], data["params"], callback, done, config);
-        case "export":
-            buff = db["export"]();
-            result = {
-                id: data["id"],
-                buffer: buff
-            };
-            try {
-                return postMessage(result, [result]);
-            } catch (error) {
-                return postMessage(result);
-            }
-        case "close":
-            if (db) {
-                db.close();
-            }
-            return postMessage({
-                id: data["id"]
-            });
-        default:
-            throw new Error("Invalid action : " + (data && data["action"]));
-    }
-}
-
-function onError(err) {
-    return postMessage({
-        id: this["data"]["id"],
-        error: err["message"]
-    });
-}
-
-// hack-sqljs - conditional-worker
-if (typeof importScripts === "function") {
-    db = null;
-    var sqlModuleReady = initSqlJs();
-    self.onmessage = function onmessage(event) {
-        return sqlModuleReady
-            .then(onModuleReady.bind(event))
-            .catch(onError.bind(event));
-    };
-}
-}
+}());
 /*jslint-enable*/
 ' >> .tmp.js
 	mv .tmp.js sqlmath_wasm.js
+)}
+
+shCiBuildWasm2() {(set -e
+# this function will build binaries in wasm
+    shCiEmsdkExport
+    # install emsdk
+    shCiEmsdkInstall
+    # cd ${EMSDK} && . ./emsdk_env.sh && cd ..
+    # build wasm
+    local OPTION
+    local FILE
+    local FILE2
+    for FILE in \
+        sqlite3.c \
+        sqlite3_ext.c \
+        sqlmath_custom.c
+    do
+        FILE2=".tmp/$(basename "$FILE").wasm.o"
+        # optimization - skip rebuild of sqlite3.c if possible
+        if [ "$FILE2" -nt "$FILE" ] && [ "$FILE" = sqlite3.c ]
+        then
+            continue
+        fi
+        OPTION=""
+        case "$FILE" in
+        sqlite3.c)
+            OPTION="$OPTION -DSQLITE3_C2"
+            ;;
+        *)
+            OPTION="$OPTION -DSQLITE3_EXT_C2"
+            ;;
+        esac
+        emcc $OPTION \
+            -DSQLITE_DISABLE_LFS \
+            -DSQLITE_ENABLE_FTS3 \
+            -DSQLITE_ENABLE_FTS3_PARENTHESIS \
+            -DSQLITE_ENABLE_MATH_FUNCTIONS \
+            -DSQLITE_ENABLE_NORMALIZE \
+            -DSQLITE_HAVE_ZLIB \
+            -DSQLITE_THREADSAFE=0 \
+            \
+            -DEMSCRIPTEN \
+            -DSQLMATH_C_DISABLE \
+            -I.tmp \
+            \
+            -O3 \
+            -Wall \
+            -flto \
+            -s INLINING_LIMIT=50 \
+            \
+            -c "$FILE" -o "$FILE2"
+    done
+	emcc \
+        -s EXPORTED_FUNCTIONS='[
+"_free",
+"_malloc",
+"_sqlite3_bind_blob",
+"_sqlite3_bind_double",
+"_sqlite3_bind_int",
+"_sqlite3_bind_parameter_index",
+"_sqlite3_bind_text",
+"_sqlite3_changes",
+"_sqlite3_clear_bindings",
+"_sqlite3_close_v2",
+"_sqlite3_column_blob",
+"_sqlite3_column_bytes",
+"_sqlite3_column_count",
+"_sqlite3_column_double",
+"_sqlite3_column_name",
+"_sqlite3_column_text",
+"_sqlite3_column_type",
+"_sqlite3_create_function_v2",
+"_sqlite3_data_count",
+"_sqlite3_errmsg",
+"_sqlite3_exec",
+"_sqlite3_finalize",
+"_sqlite3_free",
+"_sqlite3_normalized_sql",
+"_sqlite3_open",
+"_sqlite3_prepare_v2",
+"_sqlite3_reset",
+"_sqlite3_result_blob",
+"_sqlite3_result_double",
+"_sqlite3_result_error",
+"_sqlite3_result_int",
+"_sqlite3_result_int64",
+"_sqlite3_result_null",
+"_sqlite3_result_text",
+"_sqlite3_sql",
+"_sqlite3_step",
+"_sqlite3_value_blob",
+"_sqlite3_value_bytes",
+"_sqlite3_value_double",
+"_sqlite3_value_int",
+"_sqlite3_value_text",
+"_sqlite3_value_type"
+        ]' \
+        -s EXPORTED_RUNTIME_METHODS='[
+"UTF8ToString",
+"cwrap",
+"stackAlloc",
+"stackRestore",
+"stackSave"
+        ]' \
+        --memory-init-file 0 \
+        -Wall \
+        -s ALLOW_MEMORY_GROWTH=1 \
+        -s ALLOW_TABLE_GROWTH=1 \
+        -s NODEJS_CATCH_EXIT=0 \
+        -s NODEJS_CATCH_REJECTION=0 \
+        -s RESERVED_FUNCTION_POINTERS=64 \
+        -s SINGLE_FILE=0 \
+        -s USE_ZLIB \
+        -s WASM=1 \
+        \
+        .tmp/sqlite3.c.wasm.o \
+        .tmp/sqlite3_ext.c.wasm.o \
+        .tmp/sqlmath_custom.c.wasm.o \
+        \
+        -o sqlmath_wasm2.js
 )}
 
 shCiEmsdkExport() {